requirements.txt
flask
pyyaml
cryptography
reportlab
PyPDF2
pytesseract
Pillow